New filter and capapbility
--HG--
extra : convert_revision : svn%3Ac3e66f6f-232e-0410-82ea-bf51f2194cab/trunk%40690

+++ admin/album.php
@@ -91,6 +91,10 @@ class nggManageAlbum {
 		check_admin_referer('ngg_album');
 	
 		if ( isset($_POST['add']) && isset ($_POST['newalbum']) ) { 
+			
+			if (!nggGallery::current_user_can( 'NextGEN Add/Delete album' ))
+				wp_die(__('Cheatin&#8217; uh?'));			
+			
 			$newalbum = esc_attr($_POST['newalbum']);
 			$result = $wpdb->query("INSERT INTO $wpdb->nggalbum (name, sortorder) VALUES ('$newalbum','0')");
 			$this->currentID = (int) $wpdb->insert_id;
@@ -114,7 +118,10 @@ class nggManageAlbum {
 		}
 		
 		if ( isset($_POST['delete']) ) {
-	
+			
+			if (!nggGallery::current_user_can( 'NextGEN Add/Delete album' ))
+				wp_die(__('Cheatin&#8217; uh?'));
+				
 			$result = nggdb::delete_album( $this->currentID );
 			if ($result) 
 				nggGallery::show_message(__('Album deleted','nggallery'));
@@ -126,6 +133,9 @@ class nggManageAlbum {
 		global $wpdb;
 		
 		check_admin_referer('ngg_thickbox_form');
+		
+		if (!nggGallery::current_user_can( 'NextGEN Edit album settings' )) 
+			wp_die(__('Cheatin&#8217; uh?'));
 		
 		$name = esc_attr( $_POST['album_name'] );
 		$desc = esc_attr( $_POST['album_desc'] );
@@ -265,7 +275,9 @@ function showDialog() {
 				</select>
 				<?php if ($this->currentID > 0){ ?>
 					<input class="button-primary" type="submit" name="update" value="<?php _e('Update', 'nggallery'); ?>"/>
+					<?php if(nggGallery::current_user_can( 'NextGEN Edit album settings' )) { ?>
 					<input class="button-secondary" type="submit" name="showThickbox" value="<?php _e( 'Edit album', 'nggallery'); ?>" onclick="showDialog(); return false;" />
+					<?php } ?>
 					<?php if(nggGallery::current_user_can( 'NextGEN Add/Delete album' )) { ?>
 					<input class="button-secondary action "type="submit" name="delete" value="<?php _e('Delete', 'nggallery'); ?>" onclick="javascript:check=confirm('<?php _e('Delete album ?','nggallery'); ?>');if(check==false) return false;"/>
 					<?php } ?>
